backend/requirements.txt
openai
duckduckgo_search
lancedb
PyPDF2
requests
beautifulsoup4
agno
pypdf
tantivy
pandas
langchain_community
streamlit
langchain_huggingface
faiss-cpu
tiktoken
fitz
frontend
langchain_openai
flask
google-cloud-documentai